fix(overlay): play slide-out transition on card dismiss
The previous attempt placed in:fly + out:slideCollapse on
OverlayCard's outer <div>, but OverlayCard is a child component
under {#each visible as g (g.id)}. Svelte plays an outro only when
the directive sits on the immediate keyed child of the each block;
on a child component the parent each block destroys the component
synchronously and the inner transition never runs. The slide-in
appeared to work because the root element exists at mount, but
dismissal was an instant unmount with no animation.

Move the transition wiring (slideDir, reducedMotion, slideCollapse)
out of OverlayCard.svelte and into ToastFeed.svelte, wrapping
<OverlayCard> in a small <div class="tf-card-slot"> that carries
both directives. This is the immediate keyed child of the each
block, so Svelte plays both intro and outro correctly. Slide
direction is still derived from the position prop (right corners
→ +x, left → −x); reduced-motion still collapses to a short
cross-fade.

Tests updated to assert the directives live in ToastFeed.svelte
and that they precede <OverlayCard> in source order (so they sit
on the wrapper, not on the card's root).
