+++ src/lib/overlay/OverlayCard.svelte
@@ -16,79 +16,27 @@
   import {
     averageDurationMs,
     canFocusLog,
-    DEFAULT_OVERLAY_POSITION,
     groupVisualState,
     hintsForAnnotations,
     isDestructive,
     isStacked,
-    type OverlayPosition,
   } from './overlay-helpers';
   import { serverIconFor } from './icons';
   import { cardClick } from './overlay-actions';
   import HintPill from './HintPill.svelte';
   import StateIcon from './StateIcon.svelte';
-  import { fade, fly, type TransitionConfig } from 'svelte/transition';
-  import { quintOut } from 'svelte/easing';
+  import { fade } from 'svelte/transition';
 
   type Props = {
     group: ToolCallGroup;
     dismissMs: number;
     showProfile?: boolean;
-    position?: OverlayPosition;
   };
   let {
     group,
     dismissMs,
     showProfile = true,
-    position = DEFAULT_OVERLAY_POSITION,
   }: Props = $props();
-
-  // Right-anchored corners slide in/out toward +x, left-anchored toward
-  // −x. Magnitude is the card width + the feed's outer padding so the
-  // wrapper fully clears the visible area before the transition ends.
-  // 380px = --tf-card-w (340) + .tf-feed horizontal padding (20) + slack.
-  const slideDir = $derived(position.endsWith('right') ? 1 : -1);
-  const slidePx = 380;
-
-  // Honour the OS reduced-motion preference: collapse the slide to a
-  // short cross-fade with no horizontal travel. `matchMedia` is gated on
-  // `window` because vitest runs this module under node (env=node).
-  const reducedMotion =
-    typeof window !== 'undefined' &&
-    typeof window.matchMedia === 'function' &&
-    window.matchMedia('(prefers-reduced-motion: reduce)').matches;
-  const inDuration = reducedMotion ? 120 : 220;
-  const outDuration = reducedMotion ? 100 : 200;
-  const inX = $derived(reducedMotion ? 0 : slideDir * slidePx);
-  const outX = $derived(reducedMotion ? 0 : slideDir * slidePx);
-
-  // Custom out-transition: combines a horizontal fly with a height +
-  // vertical-margin collapse so the dismissing card leaves the column
-  // flow as it slides out, letting the remaining cards re-flow without
-  // a visible gap. `t` is already eased; `u = 1 - t` so the slide
-  // distance grows as the card fades.
-  function slideCollapse(
-    node: HTMLElement,
-    { x, duration }: { x: number; duration: number },
-  ): TransitionConfig {
-    const style = getComputedStyle(node);
-    const height = node.offsetHeight;
-    const marginTop = parseFloat(style.marginTop) || 0;
-    const marginBottom = parseFloat(style.marginBottom) || 0;
-    return {
-      duration,
-      easing: quintOut,
-      css: (t: number, u: number) => `
-        transform: translateX(${u * x}px);
-        opacity: ${t};
-        max-height: ${t * height}px;
-        margin-top: ${t * marginTop}px;
-        margin-bottom: ${t * marginBottom}px;
-        overflow: hidden;
-        pointer-events: none;
-      `,
-    };
-  }
 
   const state = $derived(groupVisualState(group));
   const stacked = $derived(isStacked(group));
@@ -124,8 +72,6 @@
   class="tf-card-wrap"
   data-state={state}
   data-testid="overlay-card"
-  in:fly={{ x: inX, opacity: 0, duration: inDuration, easing: quintOut }}
-  out:slideCollapse={{ x: outX, duration: outDuration }}
 >
   {#if stacked}
     <div class="tf-card tf-card-ghost tf-card-ghost-2" aria-hidden="true"></div>
